finished test for creating, retrieving and validating of user details and passwords at a database level. The ability to delete a user that **already exists** is implicit in the current testing implementation.
Further tests that need to be carried out:
is there any kind of error that is returned when a delete attempt is
made, but for an invalid user id.
is there any kind of error that is returned when there is a attempt to
retrieve a user by id, but that user actually does not exist.